analyze/requirements.txt
confluent_kafka==1.8.2
numpy==1.20.3
PyYAML==6.0
scipy==1.7.1
transformers==4.18.0
